Committing Mad Libs project
In C#, variables and string interpolation allow us to transform a piece of text by swapping out different pieces of information.

In this project, we’ll use C# to write a Mad Libs word game! Mad Libs are short stories with blanks for the player to fill in that represent different parts of speech. The end result is a really hilarious and strange story.