Defining methods
In C#, it’s convention to use PascalCase to name our method. The name starts with an uppercase letter and each word following begins with an uppercase as well. It’s not required in C#, but it makes our code easier to read for other developers.

The body of our method goes between the curly braces: { }. Whenever the method is called, the code in the body is executed.
We can also include default parameters for methods.(line 28)