Array.Find()
Array.Find() is a method that takes an array and a method as arguments. 
Array.Find() calls the method on each element of the array and returns the first element for which the method returns true.
Adjectives is an Array and method IsLong() checks for a word containing more than 8 letters. I Called Array.Find() with these two arguments to find the first element in adjectives that stands true for IsLong method's logic.